Pull out hardware functions
