+++ emulation_system/emulation_system/compose_file_creator/conversion/service_builders/abstract_service_builder.py
@@ -35,7 +35,7 @@ from emulation_system.compose_file_creator.types.intermediate_types import (
     IntermediatePorts,
     IntermediateVolumes,
 )
-from emulation_system.compose_file_creator.utilities.shared_functions import (
+from emulation_system.compose_file_creator.utilities.hardware_utils import (
     is_ot2,
     is_ot3,
 )
